# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Aman (أمان) trust-layer banking prototype API
 * OpenAPI spec version: 0.1.0
 */
import type { MonthlyStats } from './monthlyStats';
import type { ProtectionStats } from './protectionStats';
import type { Transfer } from './transfer';
import type { UserProfile } from './userProfile';

export interface DashboardSummary {
  user: UserProfile;
  protection: ProtectionStats;
  monthly: MonthlyStats;
  recentTransfers: Transfer[];
  unreadNotifications: number;
}