# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Aman (أمان) trust-layer banking prototype API
 * OpenAPI spec version: 0.1.0
 */

export interface HealthStatus {
  status: string;
}