# lib/api-zod/src/generated/types/monthlyStats.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Aman (أمان) trust-layer banking prototype API
 * OpenAPI spec version: 0.1.0
 */

export interface MonthlyStats {
  month: string;
  transferredAmount: number;
  transferCount: number;
  safeCount: number;
  flaggedCount: number;
  blockedCount: number;
}