# lib/api-zod/src/generated/types/transferDraft.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Aman (أمان) trust-layer banking prototype API
 * OpenAPI spec version: 0.1.0
 */

export interface TransferDraft {
  /** @nullable */
  beneficiaryId?: number | null;
  /** @minLength 1 */
  recipientName: string;
  /** @minLength 10 */
  recipientIban: string;
  /** @nullable */
  recipientBank?: string | null;
  /** @minimum 1 */
  amount: number;
  /** @nullable */
  purpose?: string | null;
}